components/seedgen/requirements.txt
langgraph==0.2.67
langchain-openai==0.3.2
langchain-anthropic==0.3.4
langchain_community==0.3.15
libclang==18.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
grpcio-health-checking==1.70.0
grpcio-reflection==1.70.0
grpcio-status==1.70.0
matplotlib==3.10.0
networkx==3.4.2
mypy-protobuf==3.6.0
pytest==8.3.4
jsonschema==4.23.0
pika==1.3.2
sqlalchemy==2.0.37
psycopg2-binary
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
openlit==1.33.11
redis==5.2.1
mcp-agent
anthropic
uv
mcp-server-fetch
mcp-server-tree-sitter
langchain-mcp-adapters